Carrito_project/requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
reportlab==4.4.4
sqlparse==0.5.3
tzdata==2025.2
